add tag library feature for publisher

+++ src/Tagcade/DomainManager/AdSlotManager.php
@@ -2,14 +2,20 @@
 
 namespace Tagcade\DomainManager;
 
+use InvalidArgumentException;
 use Tagcade\Exception\LogicException;
 use Tagcade\Exception\RuntimeException;
 use Tagcade\Model\Core\BaseAdSlotInterface;
+use Tagcade\Model\Core\BaseLibraryAdSlotInterface;
 use Tagcade\Model\Core\DisplayAdSlotInterface;
 use Tagcade\Model\Core\DynamicAdSlotInterface;
+use Tagcade\Model\Core\LibraryDisplayAdSlotInterface;
+use Tagcade\Model\Core\LibraryDynamicAdSlotInterface;
+use Tagcade\Model\Core\LibraryNativeAdSlotInterface;
 use Tagcade\Model\Core\NativeAdSlotInterface;
-use Tagcade\Model\User\Role\PublisherInterface;
 use Tagcade\Model\Core\SiteInterface;
+use Tagcade\Model\ModelInterface;
+use Tagcade\Model\User\Role\PublisherInterface;
 use Tagcade\Repository\Core\AdSlotRepositoryInterface;
 
 class AdSlotManager implements AdSlotManagerInterface
@@ -31,16 +37,21 @@ class AdSlotManager implements AdSlotManagerInterface
      */
     private $adSlotRepository;
 
-    public function __construct(DisplayAdSlotManagerInterface $displayAdSlotManager,
-        NativeAdSlotManagerInterface $nativeAdSlotManager,
-        DynamicAdSlotManagerInterface $dynamicAdSlotManager,
-        AdSlotRepositoryInterface $adSlotRepository
-    )
+    public function __construct(AdSlotRepositoryInterface $adSlotRepository)
     {
-        $this->displayAdSlotManager = $displayAdSlotManager;
-        $this->nativeAdSlotManager = $nativeAdSlotManager;
-        $this->dynamicAdSlotManager = $dynamicAdSlotManager;
         $this->adSlotRepository = $adSlotRepository;
+    }
+
+    public function setDisplayAdSlotManager(DisplayAdSlotManagerInterface $displayAdSlotManager) {
+        $this->displayAdSlotManager = $displayAdSlotManager;
+    }
+
+    public function setNativeAdSlotManager(NativeAdSlotManagerInterface $nativeAdSlotManager) {
+        $this->nativeAdSlotManager = $nativeAdSlotManager;
+    }
+
+    public function setDynamicAdSlotManager(DynamicAdSlotManagerInterface $dynamicAdSlotManager) {
+        $this->dynamicAdSlotManager = $dynamicAdSlotManager;
     }
 
     /**
@@ -54,18 +65,21 @@ class AdSlotManager implements AdSlotManagerInterface
     /**
      * @inheritdoc
      */
-    public function save(BaseAdSlotInterface $adSlot)
+    public function save(ModelInterface $adSlot)
     {
+        if(!$adSlot instanceof BaseAdSlotInterface) throw new InvalidArgumentException('expect BaseAdSlotInterface object');
+
         $this->getManager($adSlot)->save($adSlot);
     }
 
     /**
      * @inheritdoc
      */
-    public function delete(BaseAdSlotInterface $adSlot)
+    public function delete(ModelInterface $adSlot)
     {
-        $this->getManager($adSlot)->delete($adSlot);
+        if(!$adSlot instanceof BaseAdSlotInterface) throw new InvalidArgumentException('expect BaseAdSlotInterface object');
 
+        $this->getManager($adSlot)->delete($adSlot);
     }
 
     /**
@@ -94,7 +108,7 @@ class AdSlotManager implements AdSlotManagerInterface
 
     public function allReportableAdSlots($limit = null, $offset = null)
     {
-       return $this->adSlotRepository->allReportableAdSlots($limit, $offset);
+        return $this->adSlotRepository->allReportableAdSlots($limit, $offset);
     }
 
 
@@ -119,29 +133,45 @@ class AdSlotManager implements AdSlotManagerInterface
      */
     public function getReportableAdSlotsForPublisher(PublisherInterface $publisher, $limit = null, $offset = null)
     {
-       return $this->adSlotRepository->getReportableAdSlotsForPublisher($publisher, $limit, $offset);
+        return $this->adSlotRepository->getReportableAdSlotsForPublisher($publisher, $limit, $offset);
     }
 
     /**
-     * @param BaseAdSlotInterface $adSlot
+     * @param BaseAdSlotInterface|BaseLibraryAdSlotInterface $adSlot
      * @return DisplayAdSlotManagerInterface|NativeAdSlotManagerInterface|DynamicAdSlotManagerInterface
      */
-    protected function getManager(BaseAdSlotInterface $adSlot)
+    protected function getManager($adSlot)
     {
-        if ($adSlot instanceof DisplayAdSlotInterface) {
+        if ($adSlot instanceof DisplayAdSlotInterface || $adSlot instanceof LibraryDisplayAdSlotInterface) {
             return $this->displayAdSlotManager;
         }
 
-        if ($adSlot instanceof NativeAdSlotInterface) {
+        if ($adSlot instanceof NativeAdSlotInterface || $adSlot instanceof LibraryNativeAdSlotInterface) {
             return $this->nativeAdSlotManager;
         }
 
-        if ($adSlot instanceof DynamicAdSlotInterface) {
+        if ($adSlot instanceof DynamicAdSlotInterface || $adSlot instanceof LibraryDynamicAdSlotInterface) {
             return $this->dynamicAdSlotManager;
         }
 
         throw new LogicException('Do not support manager for this type of ad slot');
 
 
+    }
+
+    public function persistAndFlush(BaseAdSlotInterface $adSlot)
+    {
+        $this->getManager($adSlot)->persistAndFlush($adSlot);
+    }
+
+    /**
+     * Get all referenced ad slots that refer to the same library and on the same site to current slot
+     * @param BaseLibraryAdSlotInterface $libraryAdSlot
+     * @param SiteInterface $site
+     * @return BaseAdSlotInterface[]
+     */
+    public function getReferencedAdSlotsForSite(BaseLibraryAdSlotInterface $libraryAdSlot, SiteInterface $site)
+    {
+        return $this->getManager($libraryAdSlot)->getReferencedAdSlotsForSite($libraryAdSlot, $site);
     }
 }
